% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scielo_search.R
\name{scielo_search_all}
\alias{scielo_search_all}
\title{Search SciELO SOLR endpoint}
\usage{
scielo_search_all(solr_q = NULL, facet = FALSE)
}
\arguments{
\item{solr_q}{solr parameters passed on to the respective \pkg{solrium}
package function wrapped in a list.}

\item{facet}{logical, should facet with counts be returned?}
}
\value{
data.frame, or in case of `facet = TRUE` a list of all facets
}
\description{
Search SciELO SOLR endpoint
}
\details{
See the \pkg{solrium} package documentation for available
parameters.
}
\examples{
\dontrun{
# basic search
scielo_search_all(list(q = "dengue"))

# also return facets
scielo_search_all(list(q = "*:*"), facet = TRUE)
}
}
